libheif: run library init/deinit routines.
A plugin interface[1] added in libheif 1.14.0 requires running
`heif_init` to discover and load the plugins and `heif_deinit` for a
proper cleanup.
To support that, we need to extend the imv_backend structure with
init/uninit callbacks and invoke them when necessary.

Fixes discovery of libheif codecs shipped as dynamic plugins.

[1]: https://github.com/strukturag/libheif#codec-plugins

Signed-off-by: Marcin Serwin <marcin@serwin.dev>