Explicitly handle UI scale
Cairo is used both to draw UI elements and to draw the SVG pictures.
The former need to account for the compositor's scale while the latter
do not so we no longer rely on Cairo's device scale feature and instead
handle the scaling manually.

Based on the comment of ~danielrainer in the linked issue.

Fixes: https://todo.sr.ht/~exec64/imv/16
Signed-off-by: Marcin Serwin <marcin@serwin.dev>