Allow showing jpeg images with CMYK or YCCK colorspace
Links in libjpeg (provided by turbojpeg) to extract more
info from the JPEG header.

Introduces an optional new dependency (libcms2) that is used to extract
the embedded ICC profile.

If no ICC profile is present, or libcms2 is not available, a naive
conversion algorithm is used to convert the colors from CMYK to RGBA.

Another option might be to ship a default icc profile, but that is
considered overkill.

Updated the sample_cmyk.jpeg file: exporting the regular sample.jpeg
file in GIMP produced a gray pixel that was much darker than the
original. I've now done the conversion with imagemagick and it is much
closer to the original.

Reviewed-by: Marcin Serwin <marcin@serwin.dev>